Rename ConfigurationIndexation to configuration-indexation
