+++ packages/js/src/first-time-configuration/tailwind-components/steps/indexation/configuration-indexation.js
